refactor(Profile): Split DeleteAccount into two part
