+++ src/components/DeleteAccount/DeleteAccount.jsx
@@ -1,0 +1,69 @@
+import React, { useEffect, useState } from 'react'
+import { useNavigate } from 'react-router-dom'
+
+import Alerter from 'cozy-ui/transpiled/react/deprecated/Alerter'
+
+import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
+
+import ConfirmModal from 'components/DeleteAccount/ConfirmModal'
+import FormModal from 'components/DeleteAccount/FormModal'
+import EmailConfirmationModal from 'components/DeleteAccount/EmailConfirmationModal'
+import { useHasPassword } from 'hooks/useHasPassword'
+
+const CONFIRMING = 'confirming'
+const IDLE = 'idle'
+const REQUESTING = 'requesting'
+const EMAIL_CONFIRMATION = 'email_confirmation'
+
+const DeleteAccount = () => {
+  const { t } = useI18n()
+  const navigate = useNavigate()
+  const hasPassword = useHasPassword()
+
+  const [status, setStatus] = useState(IDLE)
+
+  useEffect(() => {
+    setStatus(hasPassword ? CONFIRMING : REQUESTING)
+  }, [hasPassword])
+
+  const onError = error => {
+    setStatus(IDLE)
+    console.error(error.message) // eslint-disable-line no-console
+    Alerter.error(t('DeleteAccount.error.message'))
+  }
+
+  const onRequested = () => {
+    setStatus(EMAIL_CONFIRMATION)
+  }
+
+  const handleClose = () => {
+    navigate('..')
+  }
+
+  if (status === CONFIRMING) {
+    return (
+      <ConfirmModal
+        onClose={handleClose}
+        onSuccess={() => setStatus(REQUESTING)}
+      />
+    )
+  }
+
+  if (status === REQUESTING) {
+    return (
+      <FormModal
+        onClose={handleClose}
+        onError={onError}
+        onSuccess={onRequested}
+      />
+    )
+  }
+
+  if (status === EMAIL_CONFIRMATION) {
+    return <EmailConfirmationModal onClose={handleClose} />
+  }
+
+  return null
+}
+
+export { DeleteAccount }
