+++ src/components/DeleteAccount/index.js
@@ -1,0 +1,1 @@
+export { DeleteAccount } from './DeleteAccount'
